Refactor/#4: EditModeView를 CompletedFileList 인라인 편집 모드로 통합
